# Conflicts:
#	core/server/master/src/main/java/alluxio/master/block/BlockMaster.java
#	core/server/worker/src/main/java/alluxio/worker/block/DefaultBlockWorker.java

+++ job/server/src/main/java/alluxio/master/job/JobMaster.java
@@ -175,9 +175,9 @@ public class JobMaster extends AbstractMaster implements NoopJournaled {
     mWorkflowTracker = new WorkflowTracker(this);
 
     mPlanTracker = new PlanTracker(
-        ServerConfiguration.getLong(PropertyKey.JOB_MASTER_JOB_CAPACITY),
+        ServerConfiguration.getInt(PropertyKey.JOB_MASTER_JOB_CAPACITY),
         ServerConfiguration.getMs(PropertyKey.JOB_MASTER_FINISHED_JOB_RETENTION_TIME),
-        ServerConfiguration.getLong(PropertyKey.JOB_MASTER_FINISHED_JOB_PURGE_COUNT),
+        ServerConfiguration.getInt(PropertyKey.JOB_MASTER_FINISHED_JOB_PURGE_COUNT),
         mWorkflowTracker);
 
     mWorkerHealth = new ConcurrentHashMap<>();
